+++ stories/components/data-table/advanced/data-table.advanced.stories.tsx
@@ -2,6 +2,7 @@ import { DataTable } from '@semcore/data-table';
 import type { Meta, StoryObj } from '@storybook/react-vite';
 
 import AccordionInMergedRowsExample, { accordionMergedProps } from './examples/accordion_in_merged_rows';
+import AccordionWithCheckboxExample from './examples/accordion_with_checkbox';
 import AccordionWithPaginationExample, { tableInTableDefaultProps } from './examples/accordion_with_pagination';
 import type { TableInTableProps } from './examples/accordion_with_pagination';
 import AccordionWithStickyRowsExample, { accordionStickyProps } from './examples/accordion_with_sticky_rows';
@@ -32,6 +33,10 @@ export const AccordionInMergedRows: StoryObj<typeof accordionMergedProps> = {
     loading: { control: 'boolean' },
   },
   args: accordionMergedProps,
+};
+
+export const AccordionWithCheckbox: Story = {
+  render: AccordionWithCheckboxExample,
 };
 
 export const AccordionWithPagination: StoryObj<TableInTableProps> = {