+++ semcore/data-table/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - Render table with expanded rows after changing page.
 - Render grouped header in some cases.
+- Render accordion and checkbox in one row together.
 
 ## [16.3.2] - 2025-09-20